Move fedpkg to other file
